Create Petition for peace and diplomacy
This petition gives people from Germany, Ukraine and Russia a voice. Together we say: Yes to peace, no to tax money for military purposes!

Our goal: The government should clearly commit to diplomacy and avoid military involvement.
Billions are currently flowing into the military - money that could be used for diplomacy and economic growth.

Sign now and take a stand for peace. Share the petition to reach even more people!